Add files via upload
There are too many Shared Parametre files.
This one is out here as well....
Autodesk has created files where Instance and Type Parameters are seperated, this is understandable but unpractical.
Later they added the Pset/Qset in front of them. There are 17 Psets containing FireRating, now we have 17 instance and 17 Type parameters for FireRating.

They all work.
Seperating Instance and Type is ont thing, creating a seperate Property for every Pset is antother. For Loadable Families it's fine but Project Parameters? No thanks.

This is the file I'm using. Choose what you like, there are no bad choices.
Be carefull with the GUIDS when changing parameters with the same name in your company.
Parameters are elements with GUIDS and ID's. The name is  just the human readable part. Multiple parameters with the same name can exist....